Build map from parsed accept header.
Produce a list of maps containing parts of the accept header in order to support more nuanced matching. This includes the type, subtype, quality, and extensions, as well as the implicity specificity of a media-range. The list of media-range maps is sorted using both quality and specificity.

This changes the original data structure of the value this middleware associates with the request, but given the current state of the middleware, it seems reasonable to do at this point...